requirements.txt
# MDViewer 依赖列表
# 推荐使用 requirements_no_pango.txt (无 Pango 版本)

PyQt5==5.15.9
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.7
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
pygments==2.17.2
Pillow==10.1.0
pydantic==2.5.3

# PDF 导出 - 使用 reportlab (纯 Python，无需系统库)
reportlab==4.0.4

# 注意: weasyprint 已移除，因為有版本兼容性問題
# 如需使用 weasyprint，請安裝: pip install weasyprint==60.1 pydyf==0.5.0
